+++ .env
@@ -1,3 +1,3 @@
-VITE_APP_PUBLIC_PATH=/
+VITE_APP_PUBLIC_PATH=/groupsharing
 VITE_APP_PREVIEW=true
 VITE_APP_API_BASE_URL=/api
