+++ .env.production
@@ -1,2 +1,3 @@
 VITE_APP_PREVIEW=false
+VITE_APP_PUBLIC_PATH=/groupsharing
 VITE_APP_API_BASE_URL=https://api.example.com
