+++ src/router/index.ts
@@ -3,6 +3,10 @@
 import { createRouter, createWebHistory } from 'vue-router/auto'
 import NProgress from 'nprogress'
 import 'nprogress/nprogress.css'
+import { getLoginInfo } from '@/utils/index'
+
+const LOGIN_PATH = '/PhoneLogin'
+const HOME_PATH = '/'
 
 NProgress.configure({ showSpinner: true, parent: '#app' })
 const router = createRouter({
@@ -11,7 +15,16 @@ const router = createRouter({
 
 router.beforeEach((_to, _from, next) => {
   NProgress.start() // start progress bar
-  next()
+  const loginInfo = getLoginInfo()
+  if (_to.path === LOGIN_PATH || _to.path === HOME_PATH) {
+    next()
+    return
+  }
+
+  if (loginInfo?.accessToken)
+    next()
+  else
+    next('/PhoneLogin')
 })
 
 router.afterEach(() => {
