+++ src/typed-router.d.ts
@@ -41,9 +41,9 @@ declare module 'vue-router/auto/routes' {
   export interface RouteNamedMap {
     '/': RouteRecordInfo<'/', '/', Record<never, never>, Record<never, never>>,
     '/charts/': RouteRecordInfo<'/charts/', '/charts', Record<never, never>, Record<never, never>>,
-    '/GroupSharingHome': RouteRecordInfo<'/GroupSharingHome', '/GroupSharingHome', Record<never, never>, Record<never, never>>,
     '/mock/': RouteRecordInfo<'/mock/', '/mock', Record<never, never>, Record<never, never>>,
     '/Order': RouteRecordInfo<'/Order', '/Order', Record<never, never>, Record<never, never>>,
+    '/PhoneLogin': RouteRecordInfo<'/PhoneLogin', '/PhoneLogin', Record<never, never>, Record<never, never>>,
     '/StudentInfo': RouteRecordInfo<'/StudentInfo', '/StudentInfo', Record<never, never>, Record<never, never>>,
     '/unocss/': RouteRecordInfo<'/unocss/', '/unocss', Record<never, never>, Record<never, never>>,
   }
