+++ src/views/PhoneLogin.vue
@@ -1,0 +1,13 @@
+<script setup lang="ts">
+import PageView from '@/components/PageView.vue'
+</script>
+
+<template>
+  <PageView :has-nav-bar="true" title="">
+    <div class="loginForm">
+      登录
+    </div>
+  </PageView>
+</template>
+
+<style scoped></style>
